test(comment): fixed failed test cases
