+++ app/Repositories/TaskRepository.php
@@ -2,6 +2,9 @@
 
 namespace App\Repositories;
 
+use App\Events\AddComment;
+use App\Events\DeleteComment;
+use App\Events\UpdateComment;
 use App\Models\ActivityType;
 use App\Models\Comment;
 use App\Models\Project;
@@ -50,7 +53,7 @@ class TaskRepository extends BaseRepository
     }
 
     /**
-     * @param int   $id
+     * @param int $id
      * @param array $columns
      *
      * @return Task
@@ -94,7 +97,7 @@ class TaskRepository extends BaseRepository
 
     /**
      * @param array $input
-     * @param int   $id
+     * @param int $id
      *
      * @throws Exception
      *
@@ -132,7 +135,7 @@ class TaskRepository extends BaseRepository
     }
 
     /**
-     * @param array     $input
+     * @param array $input
      * @param Task|null $task
      *
      * @return bool
@@ -227,7 +230,7 @@ class TaskRepository extends BaseRepository
     }
 
     /**
-     * @param Task  $task
+     * @param Task $task
      * @param array $tags
      */
     public function attachTags($task, $tags)
@@ -265,10 +268,12 @@ class TaskRepository extends BaseRepository
     public function getTaskDetails($id, $input = [])
     {
         if (isset($input['user_id']) && $input['user_id'] > 0) {
-            $task = Task::with(['timeEntries' => function ($query) use ($input) {
-                $query->where('time_entries.user_id', '=', $input['user_id'])
-                    ->with('user');
-            }])->findOrFail($id);
+            $task = Task::with([
+                'timeEntries' => function ($query) use ($input) {
+                    $query->where('time_entries.user_id', '=', $input['user_id'])
+                        ->with('user');
+                },
+            ])->findOrFail($id);
         } else {
             $task = Task::with('timeEntries.user')->findOrFail($id);
         }
@@ -423,5 +428,29 @@ class TaskRepository extends BaseRepository
         $comment = Comment::create($input);
 
         return Comment::with('createdUser')->findOrFail($comment->id);
+    }
+
+    /**
+     * @param Comment $comment
+     */
+    public function addCommentBroadCast($comment)
+    {
+        broadcast(new AddComment($comment))->toOthers();
+    }
+
+    /**
+     * @param Comment $comment
+     */
+    public function deleteCommentBroadCast($comment)
+    {
+        broadcast(new DeleteComment($comment))->toOthers();
+    }
+
+    /**
+     * @param Comment $comment
+     */
+    public function editCommentBroadCast($comment)
+    {
+        broadcast(new UpdateComment($comment))->toOthers();
     }
 }
